+++ sync-for-payables/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.1.0] sync-for-payables
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-sync-for-payables/0.1.0 - sync-for-payables
+
+## 2023-09-07 00:17:02
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.79.0 (2.101.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.0] sync-for-payables
+### Releases
+- [PyPI v0.2.0] https://pypi.org/project/codat-sync-for-payables/0.2.0 - sync-for-payables